+++ packages/component/src/block/shared/mobileSharePnl.tsx
@@ -16,6 +16,7 @@ import {
   useCarousel,
 } from "@/carousel/carousel";
 import { Carousel } from "@/carousel";
+import { useLocalStorage } from "@orderly.network/hooks";
 
 export const MobileSharePnLContent: FC<{
   position: any;
@@ -68,7 +69,7 @@ export const MobileSharePnLContent: FC<{
       setCarouselHeight(divHeight);
       setScale(divWidth / 552);
     }
-  }, [carouselRef]);
+  }, [carouselRef, domain]);
 
   const onSharePnL = async (
     posterRef: React.MutableRefObject<PosterRef | null>
@@ -80,7 +81,7 @@ export const MobileSharePnLContent: FC<{
       // Check if the browser supports the share feature
       if (navigator.share) {
         await navigator.share({
-          title: "Share PnL",
+          // title: "Share PnL",
           text: message,
           // url: imageUrl,
           files: [new File([blob], "image.png", { type: "image/png" })],
@@ -96,7 +97,7 @@ export const MobileSharePnLContent: FC<{
   };
 
   return (
-    <div className="orderly-p-0">
+    <div className="orderly-referral">
       <div
         ref={carouselRef}
         className="orderly-w-full orderly-mt-4 orderly-overflow-hidden"
@@ -126,8 +127,8 @@ export const MobileSharePnLContent: FC<{
           </CarouselContent>
           <div className="orderly-mt-2 orderly-mb-1 orderly-flex orderly-justify-center">
             <MyIdentifier
-              dotClassName="orderly-w-[16px] orderly-h-[4px] orderly-bg-base-300"
-              dotActiveClassName="orderly-bg-primary orderly-w-[20px]"
+              dotClassName="orderly-w-[16px] orderly-h-[4px] orderly-bg-base-300 orderly-dot-normal"
+              dotActiveClassName="orderly-bg-primary orderly-w-[20px] orderly-dot-sel"
               setSelectIndex={setSelectIndex}
             />
           </div>
@@ -216,7 +217,7 @@ export const MobileSharePnLContent: FC<{
       <div className="orderly-pt-2">
         <Button
           fullWidth
-          className="orderly-h-[40px]"
+          className="orderly-h-[40px] orderly-text-[16px]"
           onClick={() => {
             onSharePnL(posterRefs[selectIndex]);
           }}
@@ -251,9 +252,8 @@ const PnlFormatView: FC<{
   return (
     <div
       className={cn(
-        "orderly-shadow-lg orderly-rounded-lg orderly-h-[48px] orderly-flex-1 orderly-bg-base-400 hover:orderly-cursor-pointer orderly-items-center orderly-flex orderly-p-2",
-        isSelected &&
-          "orderly-outline orderly-outline-primary orderly-outline-1"
+        "orderly-shadow-lg orderly-rounded-lg orderly-h-[46px] orderly-flex-1 orderly-bg-base-400 hover:orderly-cursor-pointer orderly-flex orderly-items-center orderly-px-3 orderly-referral-shadow",
+        isSelected && "orderly-bg-primary orderly-dot-sel"
       )}
       onClick={() => {
         setPnlFormat(type);
@@ -292,7 +292,7 @@ const ShareOption: FC<{
   return (
     <div
       className={cn(
-        "orderly-shadow-lg orderly-rounded-lg orderly-h-[48px] orderly-mt-0 orderly-w-[calc(50%-6px)] orderly-bg-base-400 hover:orderly-cursor-pointer orderly-items-center orderly-flex orderly-p-3"
+        "orderly-shadow-lg orderly-rounded-lg orderly-h-[46px] orderly-mt-0 orderly-w-[calc(50%-6px)] orderly-bg-base-400 hover:orderly-cursor-pointer orderly-items-center orderly-flex orderly-p-3 orderly-referral-shadow"
       )}
       onClick={() => {
         // setPnlFormat(type);