+++ packages/hooks/src/utils/dev.ts
@@ -1,11 +1,12 @@
 import { EventEmitter, SimpleDI } from "@orderly.network/core";
+import { getGlobalObject } from "@orderly.network/utils";
 
 const __ORDERLY__DEBUGER_TAG = "__ORDERLY__DEBUGER";
 
 function debugPrint(msg: any): void;
 function debugPrint(msg: any, tag?: string) {
   tag = `${__ORDERLY__DEBUGER_TAG}:${tag} `;
-  if (!(window as any).__ORDERLY__DEBUGER) {
+  if (!(getGlobalObject() as any).__ORDERLY__DEBUGER) {
     console.log("${tag}${msg}");
     // console.info(
     //   "You can add `DebugPrinter` component to see the data in your UI"